+++ NEWS.md
@@ -1,5 +1,8 @@
 # string2path (development version)
 
+* This is a maintenance release to disable downloading pre-compiled binaries in
+  order to comply with the CRAN Repository Policy.
+
 # string2path 0.1.1
 
 * This is a maintenance release to fix some mistakes pointed by the CRAN